add -Wconversion flag for src/nvim/os/input.c + fix

+++ src/nvim/option_defs.h
@@ -588,7 +588,7 @@ EXTERN long p_ul;               /* 'undolevels' */
 EXTERN long p_ur;               /* 'undoreload' */
 EXTERN int p_unc;               /* 'unnamedclip' */
 EXTERN long p_uc;               /* 'updatecount' */
-EXTERN long p_ut;               /* 'updatetime' */
+EXTERN int32_t p_ut;            /* 'updatetime' */
 EXTERN char_u   *p_fcs;         /* 'fillchar' */
 EXTERN char_u   *p_viminfo;     /* 'viminfo' */
 EXTERN char_u   *p_vdir;        /* 'viewdir' */
